+++ secrets/consul/secret_token.go
@@ -5,6 +5,7 @@ package consul
 
 import (
 	"context"
+	"errors"
 	"fmt"
 
 	"github.com/hashicorp/consul/api"
@@ -76,7 +77,7 @@ func (b *backend) secretTokenRevoke(ctx context.Context, req *logical.Request, d
 		// nothing we can do about it. We already can't revoke the lease
 		// properly if it has been renewed and this is documented pre-0.5.3
 		// behavior with a security bulletin about it.
-		return nil, nil
+		return nil, nil //nolint:nilnil
 	}
 
 	var version string
@@ -106,18 +107,28 @@ func (b *backend) secretTokenRevoke(ctx context.Context, req *logical.Request, d
 	switch version {
 	case "":
 		// Pre 1.4 tokens
-		_, err := c.ACL().Destroy(tokenRaw.(string), nil)
+		_, err := c.ACL().Destroy(tokenRaw.(string), nil) //nolint:staticcheck
 		if err != nil {
 			return nil, err
 		}
 	case tokenPolicyType:
 		_, err := c.ACL().TokenDelete(tokenRaw.(string), revokeWriteOptions)
 		if err != nil {
+			statusError := api.StatusError{}
+
+			if errors.As(err, &statusError) &&
+				statusError.Code == 404 &&
+				// Don't just rely on the status code, a 404 could have many causes (e.g. load balancer has briefly no backend)
+				// So we additionally match the exact response body.
+				// This might break in future versions of Consul, but at least it's safe.
+				statusError.Body == "Cannot find token to delete" {
+				return nil, nil //nolint:nilnil
+			}
 			return nil, err
 		}
 	default:
-		return nil, fmt.Errorf("Invalid version string in data: %s", version)
+		return nil, fmt.Errorf("invalid version string in data: %s", version)
 	}
 
-	return nil, nil
+	return nil, nil //nolint:nilnil
 }