+++ secrets/consul/CHANGELOG.md
@@ -15,12 +15,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - plugin tests
+- ignore missing tokens during revoke
 
 ### Changed
 
 - cleanup docs
 - replace "Vault" with "OpenBao" in docs
 - use OpenBao go packages
+- refactor "main.go" to match other plugins
 
 ### Other